create S2EdgeLike.
